+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## Unrelease 1.3.0-beta.6
+### Changed
+- **ProductTile**
+    - fix servicePacks buyLink
+    - change secondParams value type
+    - add link to title
+- **ProductTileRest**
+    - change secondParams value type
+- **ProductTileOptions**
+    - change options value type
+- **TextLink**
+    - add underline style
+
 ## Release 1.3.0-beta.5
 ### Changed
 - **Carousel**